refactor: move iac related functions from output.text.message to iac.output.iac_text_output_handler
